src/features/acoustic_prosodic/acoustic_prosodic.py
"""
Acoustic and Prosodic Feature Extractor (PLACEHOLDER)

This module is a placeholder for acoustic and prosodic features to be implemented
by Team Member A. These features analyze audio characteristics of speech.

Features to be implemented:
- Pitch Features:
  - Mean pitch (F0)
  - Pitch range (min/max)
  - Pitch variability (standard deviation)
  - Pitch contour patterns
  
- Speech Rate Features:
  - Speaking rate (syllables/words per second)
  - Articulation rate
  - Pause frequency and duration
  
- Prosodic Features:
  - Intonation patterns
  - Stress patterns
  - Rhythm metrics
  - Voice quality measures
  
- Pause Patterns:
  - Silent pause duration
  - Filled pause frequency (um, uh, etc.)
  - Inter-turn pause duration

Integration Points:
- Audio files should be linked via transcript metadata
- Use libraries like: librosa, praat-parselmouth, or pyAudioAnalysis
- Extract features from WAV/MP3 files corresponding to transcripts

Author: Team Member A (To be implemented)
"""

from typing import List, Dict, Any
from src.parsers.chat_parser import TranscriptData
from ..base_features import BaseFeatureExtractor, FeatureResult


class AcousticProsodicFeatures(BaseFeatureExtractor):
    """
    PLACEHOLDER: Extract acoustic and prosodic features from audio.
    
    This class is a placeholder for future implementation.
    When implemented, it should extract audio-based features from
    recordings corresponding to the transcripts.
    
    Integration Guide for Team Member A:
    ------------------------------------
    1. Audio File Access:
       - Use transcript.metadata.get('media') to get audio file path
       - Or construct path from transcript.file_path
    
    2. Required Libraries:
       pip install librosa praat-parselmouth pyAudioAnalysis
    
    3. Example Implementation:
       ```python
       import librosa
       
       def extract(self, transcript: TranscriptData) -> FeatureResult:
           audio_path = self._get_audio_path(transcript)
           y, sr = librosa.load(audio_path)
           
           # Extract pitch
           pitches, magnitudes = librosa.piptrack(y=y, sr=sr)
           
           # Extract other features...
           
           return FeatureResult(
               features={...},
               feature_type='acoustic_prosodic'
           )
       ```
    
    4. Contact: Coordinate with main team for integration
    """
    
    @property
    def feature_names(self) -> List[str]:
        """
        Get list of acoustic/prosodic feature names.
        
        TO BE IMPLEMENTED by Team Member A.
        """
        return [
            # Pitch features (to be implemented)
            'mean_pitch',
            'pitch_std',
            'pitch_range',
            'pitch_slope',
            
            # Speech rate features (to be implemented)
            'speaking_rate',
            'articulation_rate',
            'pause_rate',
            
            # Prosodic features (to be implemented)
            'intonation_variability',
            'stress_pattern_score',
            'rhythm_score',
            
            # Pause features (to be implemented)
            'mean_pause_duration',
            'filled_pause_ratio',
        ]
    
    def extract(self, transcript: TranscriptData) -> FeatureResult:
        """
        Extract acoustic and prosodic features.
        
        PLACEHOLDER IMPLEMENTATION - Returns zeros.
        
        Args:
            transcript: Parsed transcript data with audio metadata
            
        Returns:
            FeatureResult with acoustic/prosodic features
        """
        # PLACEHOLDER: Return zero features
        # Team Member A should implement actual audio analysis here
        
        features = {name: 0.0 for name in self.feature_names}
        
        return FeatureResult(
            features=features,
            feature_type='acoustic_prosodic',
            metadata={
                'status': 'placeholder',
                'note': 'To be implemented by Team Member A',
                'audio_available': transcript.metadata.get('media') is not None
            }
        )


__all__ = ["AcousticProsodicFeatures"]

