requirements.txt
# ASD Detection from Conversational Features
# Requirements file for Phase 1 & 2 implementation

# ============================================================================
# Core Data Processing Libraries
# ============================================================================
pylangacq==0.20.0          # CHAT/CHILDES file parsing and analysis
pandas==2.1.3              # Data manipulation and analysis
numpy==1.26.2              # Numerical computing
scipy==1.11.4              # Scientific computing

# ============================================================================
# Natural Language Processing
# ============================================================================
nltk==3.8.1                # Natural Language Toolkit
spacy==3.7.2               # Advanced NLP library
textstat==0.7.3            # Text statistics and readability metrics

# ============================================================================
# Machine Learning (for Phase 3+)
# ============================================================================
scikit-learn==1.3.2        # Machine learning algorithms
xgboost==2.0.2             # Gradient boosting
lightgbm==4.1.0            # Light gradient boosting
imbalanced-learn==0.11.0   # Handling imbalanced datasets

# ============================================================================
# Utilities
# ============================================================================
python-dotenv==1.0.0       # Environment variable management
loguru==0.7.2              # Advanced logging
pydantic==2.5.2            # Data validation
tqdm==4.66.1               # Progress bars
openpyxl==3.1.2            # Excel file reading
python-docx==1.1.0         # Word document reading

# ============================================================================
# Development & Testing
# ============================================================================
pytest==7.4.3              # Testing framework
pytest-cov==4.1.0          # Coverage reporting
black==23.12.0             # Code formatting
flake8==6.1.0              # Linting
mypy==1.7.1                # Static type checking

# ============================================================================
# Data Visualization (Optional for analysis)
# ============================================================================
matplotlib==3.8.2          # Plotting library
seaborn==0.13.0            # Statistical visualization
plotly==5.18.0             # Interactive plots

